Creative Celebrations App
This app was created for the purpose of a business competition. I was
competing in the category of small business team in the club Business
Professionals of America. We were acting as a consulting company to an
event planner. I created an app based on this said company and made
improvements. The app is not fully functional and the client database
is empty because it is not real.